# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Piano Shop Api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/product_create": {
            "post": {
                "tags": [
                    "Product Api"
                ],
                "description": "use for insert product to database",
                "operationId": "c696c9f78a1c3a849b30f431f3397634",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of the product",
                                        "type": "string"
                                    },
                                    "slug": {
                                        "description": "Slug of the product",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price of the product",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "reviews": {
                                        "description": "Reviews of the product",
                                        "type": "integer"
                                    },
                                    "rating": {
                                        "description": "Rating of the product",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "stock": {
                                        "description": "Stock of the product",
                                        "type": "integer"
                                    },
                                    "image": {
                                        "description": "Image of the product",
                                        "type": "string"
                                    },
                                    "discount": {
                                        "description": "Discount of the product",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "description": {
                                        "description": "Description of the product",
                                        "type": "text"
                                    },
                                    "status": {
                                        "description": "Status of the product",
                                        "type": "string"
                                    },
                                    "brand_id": {
                                        "description": "Brand id of the product",
                                        "type": "integer"
                                    },
                                    "category_id": {
                                        "description": "Category id of the product",
                                        "type": "integer"
                                    },
                                    "color_id": {
                                        "type": "array",
                                        "items": {
                                            "description": "Color id of the product",
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/v1/product_update/{id}": {
            "put": {
                "tags": [
                    "Product Api"
                ],
                "description": "use for update product to database",
                "operationId": "b619a0ef694283a44d4c61e47bc53f8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of the product",
                                        "type": "string"
                                    },
                                    "slug": {
                                        "description": "Slug of the product",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price of the product",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "reviews": {
                                        "description": "Reviews of the product",
                                        "type": "integer"
                                    },
                                    "rating": {
                                        "description": "Rating of the product",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "stock": {
                                        "description": "Stock of the product",
                                        "type": "integer"
                                    },
                                    "image": {
                                        "description": "Image of the product",
                                        "type": "string"
                                    },
                                    "discount": {
                                        "description": "Discount of the product",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "description": {
                                        "description": "Description of the product",
                                        "type": "text"
                                    },
                                    "status": {
                                        "description": "Status of the product",
                                        "type": "string"
                                    },
                                    "brand_id": {
                                        "description": "Brand id of the product",
                                        "type": "integer"
                                    },
                                    "category_id": {
                                        "description": "Category id of the product",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/product_delete/{id}": {
            "delete": {
                "tags": [
                    "Product Api"
                ],
                "description": "use for delete product from database",
                "operationId": "93ee1544fc71de04cbaff9e71d998aa7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/product_show_by_slug/{slug}": {
            "get": {
                "tags": [
                    "Product Api"
                ],
                "description": "use for show product by slug from database",
                "operationId": "96dda50d683f97694ec8392e3d0cd8eb",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/product_show_by_id/{id}": {
            "get": {
                "tags": [
                    "Product Api"
                ],
                "description": "use for show product by id from database",
                "operationId": "013eecc72e12b9d1cf763b23be8818ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/product_update_stock/{id}": {
            "put": {
                "tags": [
                    "Product Api"
                ],
                "description": "use for update stock of product",
                "operationId": "799b8f7342ab7e1ad0c395b363d6e0df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "stock": {
                                        "description": "Stock of the product",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    }
}